Rename: loggin check하는 함수 middleware로 이동
- 요청 시 jwt 토큰을 g에 넣어주는 미들웨어를 추가하였습니다.
- login_required에서 useless import를 제거하였습니다.
