+++ app/__init__.py
@@ -2,6 +2,9 @@ import os
 from flask import Flask
 from dotenv import load_dotenv
 
+from app.db import get_db
+from app.middleware.load_logged_in_user import load_logged_in_user
+
 load_dotenv(verbose=True)
 
 
@@ -26,13 +29,15 @@ def create_app(test_config=None):
     except OSError:
         pass
 
+    from .middleware import load_logged_in_user
+    app.register_blueprint(load_logged_in_user.bp)
+
     # 이곳에 컨트롤러 모듈관리를 해주세요
     from .controller import ctr_test
     app.register_blueprint(ctr_test.bp)
 
     from .controller import account
     app.register_blueprint(account.bp)
-
 
     app.debug = True
     return app
